requirements.txt
PySide6>=6.5.1
Pillow>=5.2.0
psutil>=5.9.5
requests>=2.31.0
python-slugify>=1.2.1
raven>=6.0.0
mozjpeg_lossless_optimization>=1.1.2
natsort[fast]>=8.4.0
distro>=1.8.0
rarfile==3.0
patool